Fix .field with single field to not wrap in array
The FieldFilter with a single field was wrapping the field value in an
array due to the degeneration of the multiple-assignment syntax.  In a
single field case, it assigned the array of fields to a single variable,
which neglected to destructure the array of fields.

Add a special case when only one field is declared.