Allow search() to use fulltext search on MySQL
Search is slow on large tables (3M rails requests, 5GB) using the SQL
"like" operator.

Add a fragile optimization to allow use of a fulltext search index on
MySQL.  The optimization is fragile because it doesn't currently check
the actual storage engine; it just checks that the connection is a MySQL
connection.  It also doesn't have spec coverage... yet.