Optimize file reading by avoiding allocating Event objects
All stages of the common Rails log parsing pipeline are slow.
Currently, we are creating an Event object for every line of the
file, which hits the memory allocator heavily.

It would be nice to have a uniform superclass for everything in the
pipeline, but at least String conforms to the to_s part of the API.  The
text accessor is redundant with that.  Event offers little more.

Tolerate a naked String instead of wrapping it in an Event in the
FileSource and also the less-heavily encountered
AppendIndentedFilter#merge_events.

The FileSource optimization increases throughput of the file source by
50% on 1.8.7 and 1.9.2 and doesn't break any specs.  The append_indented
change has no noticeable affect on my sample file, which has very few
multi-line log entries.