Avoid prefix method call in append_indented
All stages of the common Rails log parsing pipeline are slow.
Currently, we are calling the prefix method of append_indented and
likely creating a new Ruby string (even if the underlying char* is
reused).  Memory allocations are slow.

Cache the result of the prefix, which should not need to change mid-way
through stream.

This increases throughput of the append_indented filter by 60%.  on
1.9.2 and doesn't change outside behavior so long as the prefix method
is not overridden.