Optimize the field filter by creating Arrays instead of Events
The field filter is the slowest one, and it gets quite slow the more
fields are parsed.  (This is related in part to the complexity of the
regex, not just the number of fields per se.)

We currently create instances of an event subclass that holds an
instance variable for each field.  While this is a natural Ruby
implementation, it means allocating an object plus an array for
every event.

Instead, create an instance of a subclass of Array and settle for
marginally slower field accesses.  (How much slower? Don't know yet.)
Field are accessed by methods that index the array.

It would be nice to have a uniform superclass for everything in the
pipeline, but at least our Array subclass responds to the right field
names, to_s, and text methods.

This increases throughput of the field filter from X to Y on 1.9.2

Also add file and line number to class_eval call.